Changes the speeds of the vehicles and the mali for the cars.

+++ similar2logo-examples/src/main/java/fr/lgi2a/similar2logo/examples/transport/model/agents/CarFactory.java
@@ -93,7 +93,7 @@ public class CarFactory {
  			double initialAcceleration,
  			double initialX,
  			double initialY,
- 			int maxSpeed,
+ 			double maxSpeed,
  			int maxCapacity) {
 		if( ! category.isA(TurtleAgentCategory.CATEGORY) ) {
  			throw new IllegalArgumentException( "Only turtle agents are accepted." );
